refactor(Explorer): move prepareExplorer function where it makes more sense
